feat(Yeaabsira.user) add authentication

+++ Backend/starter_project_zsilentcoders/BlogApp3/BlogApp.API/Controllers/AuthController.cs
@@ -8,13 +8,13 @@ using BlogApp.Application.Features.Users.DTOs;
 using BlogApp.Application.Responses;
 using MediatR;
 using AutoMapper;
-
+using BlogApp.Api.Controllers;
 
 namespace BlogApp.API.Controllers;
 
 [ApiController]
 [Route("api/[controller]")]
-public class AuthController : ControllerBase
+public class AuthController : BaseController
 {
     private readonly IAuthService _authService;
     private readonly IMediator _mediator;
@@ -32,37 +32,75 @@ public class AuthController : ControllerBase
     public async Task<ActionResult<LoginResponse>> Login([FromBody] LoginModel request)
     {
         var response = await _authService.Login(request);
-        return Ok(response);
+        return HandleResult(response);
         
     }
 
     [HttpPost]
     [Route("Register")]
-    public async Task<ActionResult<RegistrationResponse>> Register([FromBody] RegisterDto registerDto)
+    public async Task<ActionResult<Result<RegistrationResponse>>> Register([FromBody] RegisterDto registerDto)
     {
-        var registrationResponse = await _authService.Register(_mapper.Map<RegistrationModel>(registerDto));
+        var response = await _authService.Register(_mapper.Map<RegistrationModel>(registerDto));
         var command = new Create_UserCommand {Create_UserDto = _mapper.Map<Create_UserDto>(registerDto)};
-        command.Create_UserDto.AccountId = registrationResponse.Id;
+        if (!response.Success || response.Value == null)
+            return HandleResult(response);
+
+        command.Create_UserDto.AccountId = response.Value.UserId;
         try
         {
-            var response = await _mediator.Send(command);
-            return Ok(response);
+            var userResponse = await _mediator.Send(command);
+            return HandleResult(response);
         }
         catch(Exception e)
         {
             await _authService.DeleteUser(registerDto.Email);
-            return StatusCode(500, new { Message = e.Message });
-
+            response.Success = false;
+            response.Errors.Add(e.Message);
+            return HandleResult(response);
         }
         
     }
 
     [HttpGet]
+    [Route("ResendConfirmLink")]
+    public async Task<ActionResult<Result<string>>> ResendConfirmEmailLink(string email)
+    {
+        var response = await _authService.sendConfirmEmailLink(email);
+        return HandleResult(response);
+    }
+
+    [HttpGet]
     [Route("Confirm")]
-    public async Task<ActionResult<ConfirmEmailResponse>> ConfirmEmail(string token, string email)
+    public async Task<ActionResult<Result<string>>> ConfirmEmail(string email, string token)
     {
         var response = await _authService.ConfirmEmail(token, email);
-        return Ok(response);
+        return HandleResult(response);
         
+    }
+
+    [HttpGet]
+    [Route("ForgotPassword")]
+    public async Task<ActionResult<Result<string>>> ForgotPassword(string email)
+    {
+        var response = await _authService.ForgotPassword(email);
+        return HandleResult(response);
+        
+    }
+
+    [HttpPost]
+    [Route("ResetPassword")]
+    public async Task<ActionResult<Result<string>>> ResetPassword([FromBody] ResetPasswordModel resetPasswordModel)
+    {
+        var response = await _authService.ResetPassword(resetPasswordModel);
+        return HandleResult(response);
+        
+    }
+
+    [HttpDelete]
+    [Route("Delete")]
+    public async Task<ActionResult<bool>> Delete(string email)
+    {
+        var response = await _authService.DeleteUser(email);
+        return Ok(response);  
     }
 }
